chore: added in-app event listeners (#180)
