+++ lib/customer_io.dart
@@ -5,7 +5,6 @@ import 'package:flutter/cupertino.dart';
 
 import 'customer_io_config.dart';
 import 'customer_io_enums.dart';
-import 'customer_io_inapp.dart';
 import 'customer_io_platform_interface.dart';
 import 'extensions/map_extensions.dart';
 import 'messaging_in_app/platform_interface.dart';
@@ -161,16 +160,5 @@ class CustomerIO {
   void setProfileAttributes({required Map<String, dynamic> attributes}) {
     return _platform.setProfileAttributes(
         attributes: attributes.excludeNullValues());
-  }
-
-  /// Subscribes to an in-app event listener.
-  ///
-  /// [onEvent] - A callback function that will be called every time an in-app event occurs.
-  /// The callback returns [InAppEvent].
-  ///
-  /// Returns a [StreamSubscription] that can be used to subscribe/unsubscribe from the event listener.
-  StreamSubscription subscribeToInAppEventListener(
-      void Function(InAppEvent) onEvent) {
-    return _platform.subscribeToInAppEventListener(onEvent);
   }
 }
